ssh: Wait for access to propagate through AWS
Currently, there is a period of time between which access is configured in AWS
and when access is available for use. Previously, we'd fail with an AccessDeniedException
if we attempted to start the SSH session within this time interval.

This commit polls for access, and does not start the SSH session until access is
available for use.