ssh: Forward TTY to ssm
Previously, I couldn't figure out how to forward the calling TTY to the spawned
`aws ssm` process. This is because I was manually piping stdio from the calling TTY
to the spawned process's io. Node provides no mechanism to enabled a spawned
process's io to operate as a TTY. This causes:
1. A potential lack of proper interpretation of ANSI escape sequences.
2. Inability to size / resize the resulting display.

I had been working around this by shimming the process with a pty (pseudoterminal)
implementation. However, this has two drawbacks:
1. The pty bindings require native bindings to be compiled on installation,
   reducing the portability of the p0 CLI.
2. Pseudoterminals only support a single output stream, meaning a caller could
   not separate output to stdout vs. stderr. This meant that the output of
   `p0 ssh` could not be properly piped to another process.

This commit solves all of the above by _replacing_ the spawned process's stdio
with the p0 CLI's stdio directly, using the `'inherit'` option on node's
`child_process.spawn` function.