Send a body request when assuming roles with SAML (#87)
This PR changes the request for assuming roles from using `GET` to
`POST`.

Our SAML Assertion will be too long for the encoded url search params if
we are assuming too many roles. Instead we can send a `POST` message
with the request in the body to get around length requirements.