Basic gcloud ssh (#90)
Support ssh with gcloud using native ssh with public key authentication
and ProxyCommand - the same way as the existing AWS ssh works.

The underlying ProxyCommand in gcloud's case is `gcloud compute
start-iap-tunnel`. The IAP tunnel is necessary for accessing VMs without
an external IP.

All of the existing SSH functionality is supported: `-A`, `-N`, `-L`,
`--sudo`, and `scp`.

However, login to Google Cloud is separate from `p0 login`. There is no
support for automatic log in to Google Cloud with SAML (or other way).
If we detect that the issue is gcloud login we print `Please login to
Google Cloud CLI with 'gcloud auth login'`

- Rename variables and rewrite comments to apply to both AWS and GCP
- Rework types to more closely resemble backend types and share common
fields
- Add detection gcloud-specific access propagation detection