Pre-test for sudo in gcloud ssh (#95)
Previously, we were not able to verify if sudo access has propagated to
Google Cloud VMs, specifically in the case when the user already had
regular access and then requested sudo access. This is because the IAM
permissions for regular access already exist, so we successfully logged
in the user to the VM. However, the new IAM permissions to allow sudo
have not propagated yet, so the user was not able to successfully
execute any sudo commands until later.

This PR addresses the issue by executing a pre-test ssh command that
runs `sudo -v` in non-interactive mode to verify that the user _can_
sudo. Once the test completes we go on to log the user in.

Also refactors code to move aws and gcloud checks from
`plugins/ssh/index.ts`.

Completes ENG-2292