Automatic gcloud CLI installation (#99)
Add a new `ensureInstall` method to the SSH provider type. Move the AWS
SSM installation to this function, and add a new implementation for GCP
that installs the gcloud CLI.

The `ensureInstall` method must be called before any other functions on
the SSH provider because those functions may already use the required
dependencies. In the case of GCP, the `toCliRequest` function calls
`gcloud auth print-access-token`. Add a new `prepareRequest` function is
to establish the correct ordering of calls.


https://github.com/user-attachments/assets/ed96f7f4-cc34-4563-9683-c43cb21900c1

Also: logging improvement in access propagation guard: log first
attempt, and log whether it's a pre-test (used only by gcp) or actual
ssh session login.