ENG-2449 Add error handling for importSshPublicKey call (#104)
The CLI sometimes fails with this error:

```
TypeError: Cannot read properties of undefined (reading 'posixAccounts')
    at /opt/homebrew/lib/node_modules/@p0security/cli/dist/plugins/google/ssh-key.js:69:40
    at Generator.next (<anonymous>)
    at fulfilled (/opt/homebrew/lib/node_modules/@p0security/cli/dist/plugins/google/ssh-key.js:5:58)
    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)
```

It seems related to gcloud login, even though "gcloud auth
print-access-token" either prints a token or otherwise returns with exit
code 0.

It is possible that there is a corner case here where gcloud returns an
expired access token.

Since exit code 0 is returned, we continue to upload the SSH public key,
which then silently fails, leading to the subsequent error.

Fix: Validate the response from the importSshPublicKey call and verify
that it returns HTTP status 200 OK.

Note that I have been unable to reproduce this issue. This is my
understanding from a read-through of the code qnd the error message.