[ENG-2790] Fix broken p0 kubeconfig command (#129)
This PR unbreaks the `p0 kubeconfig` command. Some values and structures
got changed around in the k8s integration config and responses from the
backend:
- Required AWS IDC metadata is no longer found in the `eksGenerated`
object; it's now in `awsResourcePermission`
- AWS account ID is no longer stored separately in the k8s integration
configuration, and now is extracted from the EKS cluster ARN
- `provider` was renamed `hosting` in the k8s integration configuration

It also adds some minor quality-of-live improvements to the `p0
kubeconfig` command, such as utilizing `spinUntil()` for some of the
longer `await`s, and adds a more robust and reusable simple AWS ARN
parser.