Add ssh-keygen command (#133)
This command can be used if user only wants to create the key pair for
SSH. One use case when user requests access through Slack instead of
through the CLI and they need to provide the public key in Slack.

The command could be extended to force regeneration of the key pair. I
did not include that option because already provisioned keys would not
work anymore.