SSH requesting access message (#137)
This PR makes two changes to the status messages we use for SSH requests
in the CLI.

1. Replaces the "Requesting access" status message with a more generic
"Checking for access in P0"
2. Adds the message "Existing access found. Connecting to node." in the
scenario where a user had access approved previously and is waiting to
be connected to the node. Previously we showed nothing.

This only affects `p0 ssh <destination>` and `p0 scp <destination>`. It
does not affect any `p0 request` flows.

Before:


https://github.com/user-attachments/assets/cc0fce18-ab55-4074-af16-14a43a6ce06f


After:


https://github.com/user-attachments/assets/3c31ae41-437f-4328-a798-096710515dcc



Addresses:
https://linear.app/p0-security/issue/ENG-2827/requesting-access-spinner-is-displayed-when-it-shouldnt

---------

Co-authored-by: Gergely Dányi <gergas3@gmail.com>