ls: display number of items that were truncated  (#141)
Addresses ENG-2998. Improves CLI usability when there are a large number
of results by giving the user an idea of the total number of results for
their query. We request twice the requested size, and if the total
number of results is within the amount, we display the number of
remaining results. If not, we just display 'of many'

<img width="737" alt="image"
src="https://github.com/user-attachments/assets/8ecf9774-0b83-4599-b398-00f6f9191759">
<img width="826" alt="image"
src="https://github.com/user-attachments/assets/74225400-ba74-4d53-b737-d16b4b3dcd03">

---------

Co-authored-by: Nathan Brahms <nbrahms@users.noreply.github.com>