Add debugging output for Azure SSH provider (#145)
This PR improves the Azure SSH debugging experience by adding debug
output for the various `az` commands being run before opening the
`ssh`/`scp` session. Specifically, when the `--debug` flag is set, the
CLI will output the stdout and stderr for all the `az` commands it runs.
Note though that it _won't_ set the `--debug` flag for the `az` commands
themselves as this outputs far too much information, with the exception
of the `az network bastion tunnel` command; while it still outputs far
too much information in debug mode, it has an unfortunate tendency to
hide certain errors unless you pass it the `--debug` flag, so I did add
it there.

Some additional information is also logged by the CLI itself now when
the `--debug` flag is set.