Add sudo propagation check for azure (#147)
This PR implements Azure's `preTestAccessPropagationArgs` and adds a new
parameter to the `SSHProvider` type to support valid access errors.

Sudo access to an azure node can sometimes take an unexpectedly long
time to propagate fully and if a user connects before sudo access
propagates then they'll need to restart their session in order to pick
up the sudo access. Therefore it is necessary to validate sudo access
before letting a user complete their connection.

We get an error code of 1 instead of an error code 0 after sudo access
is granted when performing the pretest for an Azure instance. This means
that we needed to extend the `SSHProvider` interface a little to include
potentially valid error messages and override the error code of 1.

> In the case a user doesn't have sudo access, sudo -nv will exit with
code 1 and print Sorry, user ${username} may not run sudo on ${host}. to
stderr.
>
> If the user does have sudo access, sudo -nv will exit with code 1 (so,
the same as the no access case) but print sudo: a password is required
to stderr.