Azure SSH: clear current identity before logging in again (#149)
This PR adds runs the `az logout` before logging in so that we wipe any
currently connected identities. This resolves a handful of issues we're
seeing for example:

1. granting a user access, relinquishing access, and then granting again
will use an old cached identity which will have a temporary block on
access.
2. granting sudo access, and regular user access and relinquishing one
temporarily blocks access to the account.
3. if someone runs `az login` and uses a different identity that has the
same subscription id.