Allow ctrl+d to exit terminal when valid access patterns exist (#150)
We recently introduced `validAccessPatterns` in #147 which checks to
make sure a certain termination string exists when performing
propagation checks. Consequentially for Azure, when we exit the terminal
we're now always attempting to verify that an ephemeral access message
is valid and failing the check.

Two fixes were possible, 1) we either only check `isValidError` if we're
in a propagation check or 2) we default `isValidError` to true and we
only set it to false if we encounter an error.

We didn't go with the first option in #147 because we want to remove the
concept of `isAccessPropagationPreTest` from the
`accessPropagationGuard`, but checking for valid error messages should
only apply during the pretest phase. This PR instead only passes
`validAccessPatterns` if we're performing a pre-test to the propagation
guard.