Azure SSH: provide user friendly error messages (#152)
This PR applies a small refactor that improves the error handling of
during auth steps for the Azure SSH plugin. When a user cancels their
login attempt or if we fail to set the users active Azure subscription.
The refactor introduces a knew `KnownError` type and a utility method
`normalizeAzureCliError` that should make it easy to extend and capture
new errors as we discover them.