Azure SSH: use generated linuxUserName from backend service (#154)
Only the azure principal name can be used as the linux identity. This
change makes sure that we always default to the generated linuxUserName
when attempting to log into an SSH node using Azure. The user principal
in the permission points to the user's email in P0 which can be
different than the identity they have in Azure Active Directory.

If the user accidentally signs in with the wrong account they will get
an error that the expected user name doesn't match their current
identity. This makes it easier to debug `public key` and `permission
denied` errors.