Make Google OAuth client parameters optional (#156)
The `google` object in the tenant config contains the Client ID and
Client Secret.
This config is then stored in plain text in the user's home directory.
Although the Client Secret is not really a secret, storing it in plain
text gives the impression that the P0 CLI is not secure.

This PR renames the `clientSecret` to `publicClientSecretForPkce` and
adds more explanation in code comments. (Although these comments are not
visible in the saved `config.json` file.)

The PR also makes the `google` parameters optional so this becomes a
non-issue for organizations with different login methods like Okta,
Ping, etc.