Validate tenant access (#163)
This PR adds an additional check to the frontend that will automatically
log a user out if they aren't authorized to access that tenant or that
tenant doesn't exist. This helps prevent a poor user experience if the
user mistyped the org Id or tried to access an org they are no longer a
part of.