ssh-resolve and ssh-proxy (#161)
This is to allow users to run
`ssh <node-name>` rather than `p0 ssh <node-name>`

It adds two commands to the cli, `ssh-resolve` and `ssh-proxy` meant to
be used in an ssh config such as the following
```
Match exec "p0 ssh-resolve %h -q"
 Include .p0/ssh/configs/*.config
```

`ssh-resolve` checks if the resource exists in p0 and if the user has
`approved` access to it. If so it sets up any needed keys and writes an
ssh config to `.p0/ssh/configs/` as well as request information to a
temporary file, and returns 0 (success). Otherwise it returns 1
(silently if `--quiet` is passed)

`ssh-proxy` is invoked as the `ProxyCommand` by `ssh` when the config
file generated by `ssh-resolve` is used, and reads the files written by
`ssh-resolve` (and passed to it as cli args) and runs the relevant proxy
command as well as any needed setup/teardown

---------

Co-authored-by: Brendan Berrigan <brendan.berrigan@p0.dev>