Azure SSH: reduce the number of logins (#159)
# Overview
This PR reduces the number of login attempts for Azure. Instead of
logging out and logging after every single access request, we now only
attempt a login when:

1. A user has been granted access to a subscription they didn't have
access to previously.

And we provide a verbose error message
2. A user's token issuer has become invalid.
3. A user's token cache has become invalid (they logged out or changed
accounts unexpectedly)

# Understanding the different authentication failure modes

## A user has gained access to a subscription they didn't have access to
previously.
This happens more frequently to guests, most orgs will have members that
are assigned to a tenant/subscription which prevents this error from
coming up. There's no way for us to poll for a subscription without
explicitly logging out and logging in again. The user is warned that
they may need to try again in a few minutes.

```
Subscription bcf379ca-fdd9-4ac4-adae-8f33c376fc6c not found. If access was recently granted, please try again in a few minutes. If the issue persists, please contact support@p0.dev.
```

## A user's token issuer has become invalid.
In order to support the second use case above I had to add an abort
controller and listen to messages from the azure bastion tunnel telling
us the access is invalid. Instead of trying to automatically log a user
in we'll tell them to execute a few commands to refresh their azure
cache. If they don't like that this is being done explicitly we can
refactor this code to automatically refresh their azure cache and try
again. As a first step I think it makes sense not to do things
implicitly, the failure modes get more convoluted and the code gets
significantly more complex. This error is rare and should only occur if
a user's subscriptions/management groups are changing frequently.

## A user's token cache has become invalid.
Same as above except this happens when a user logs out while trying to
access a virtual machine.

# Removing known errors
This PR removes the `KnownError` type and it's corresponding function
because it did not provide enough flexibility for processing errors.