Support ARNs as host names in the ssh-resolve command (#182)
AWS SSM managed instance ARNs are of the form:
```
arn:aws:ssm:us-west-1:391052057035:managed-instance/i-024f03df9d0417bc8
```

The forward slash in this pattern causes an issue when we try to create
the file that holds the ssh config for this instance:

This change sanitizes the host name. It goes beyond the minimal fix to
be more defensive by also removing other potentially invalid characters.