Handle "reason required" error (#184)
This PR contains two changes that are interconnected, and should
accompany a change in the P0 backend at the same time.

1. Update the name of the `--approved-only` CLI argument to `approved`.
This change is required to match the same rename in the P0 backend.

The `--approved` flag is used in the ssh-resolve command. The change in
the P0 backend modifies which instances are returned as `--approved`
(which instances are eligible for direct `ssh`). Before, only
pre-approved and pre-existing access was returned. Now persistent
("Always allow" aka auto-approved) access is also returned. That leads
to an issue with the `reason` argument because users cannot provide it
in vanilla `ssh`.

2. Allow the `reason` argument to be specified via the `P0_SSH_REASON`
env var. Catch the "reason required" error in ssh-resolve and print a
helpful error messsage.