Single-artifact CLI builds (#186)
**Problem**

Some customers are interested in a CLI build that can be installed
without having to install NodeJS or any of our other dependencies
independently, to reduce the supply-chain attack surface area.

**Change**

Adds some scripts to generate a CLI artifact using NodeJS's experimental
single executable application functionality. See
https://nodejs.org/docs/latest-v20.x/api/single-executable-applications.html

**Validation**

To build:
- Run `yarn build` to compile typescript
- From the root directory, run `source build.sh` to generate the CLI
artifact

To test binary:
- Run `nvm deactivate` to temporarily remove NodeJS from your path
- Ensure that `which node` indicates that node cannot be found
- Run the p0 artifact to login and make a request

To test install via NPM:
- Remove global npm install
- Run `yarn build` && `npm pack`
- Run `npm install -g <npm_packed_file>`
- Run the p0 command to login and make a request

**Tracking**


https://linear.app/p0-security/issue/ENG-4341/package-p0-cli-into-a-single-deployable-artifact

**Implementation**

- Moves all build artifacts, including the tsc output, to a `/build`
directory.
- The `source build.sh` command is outside of the existing `yarn build`
script, because for some reason it wouldn't work if I included in there.
- The build is currently only tested on MacOS. Some of the build
commands might need to be modified for Windows and Linux (see the NodeJS
documentation for details)
- As-is, the CLI artifact generated by these scripts may not work if you
move it to a different device, due to code signing.