Password login (#192)
Add support for email / password login to P0 backend.

The email and password are provided by the user via the `P0_EMAIL` and
`P0_PASSWORD` environment variables to avoid storing these in the cached
identity file.

If an org uses password login then their "login" is the same as
authenticating to firebase. That, however is not deduplicated.
`loadCredentialsWithAutoLogin` and `authenticateToFirebase` both sign in
the user with email / password to firebase. Although this is redundant,
keeping that logic makes the change much smaller and easier to see that
existing functionality is not affected.