Alias generated kubeconfig entries (#199)
Previously, the generated kubeconfig user and generated kubeconfig
context defaulted to the cluster ARN. If the user had already configured
the cluster then the `aws eks update-kubeconfig` command overwrote the
existing context and user, adding the AWS_PROFILE env var to the user.
Once access expired this env var remained in the kubeconfig, essentially
making the context unusable.

This change aliases the context name and the user name by prefixing them
with `p0cli-managed-eks`, so P0 updates its own entries instead of
modifying existing ones.

Added error detection for the case when the `--user-alias` option is not
supported in the user's aws CLI because it's at an old version.