Allow main to be run if isSea() is true (#203)
It looks like our current GHA builds of the CLI are essentially no-ops.
I discovered this when trying to debug the .pkg build. I'm not sure why
I didn't see this before, perhaps it had something to do with how I was
building and/or testing locally.

Validated by running the SEA build locally and testing the binary can
make requests.