General improvements to CLI version management (#206)
Running `p0 --version` using the standalone CLI prints "unknown". This
is likely due to yargs' default --version parser not being able to
locate package.json. That + a few other fixes.

- Makes it so the version in package.json is the source of truth (no
need to change it in multiple places)
- Add a function that identifies the version in all environments
- Use this implementation everywhere we need the version number
- Fix a potential bug where the auto-update message may not display
- Improve docs and messaging indicating how to install / upgrade

Tested by building locally and seeing that the version prints itself and
the CLI still runs