OpenTelemetry tracing (#211)
Add OpenTelemetry tracing and create a root span in `main`.

- The collector endpoint is an endpoint in the P0 backend, that is
authenticated. This means we can only export spans _after_ the user has
logged in.
- Add a custom `BufferedSpanExporter` exporter that just buffers spans
in memory until login is complete and JWT is available to send in the
header. In practice, we are not actually buffering spans yet since
export is not triggered until after login is complete since we don't
have enough spans yet. In practice export is triggered forcefully before
CLI exits. Note that even if we omitted the buffering functionality we'd
still need something like the `BufferedSpanExporter` that allows
injecting the delegate `OTLPTraceExporter` after authentication is
complete, because the configuration of the `OTLPTraceExporter` cannot be
changed dynamically, and we have to change it - put the JWT in the
header.
- Change version loading to be synchronous, in order to use it in
`instrumentation.ts` for annotating spans with the CLI version. We
cannot have top-level awaits.
- `src/index.ts` is ignored from prettier check because the import
ordering doesn't comply. We first import the instrumentation, start the
tracer, and then import the rest. This is important so OpenTelemetry can
auto-instrument all libraries.