ENG-5460 remove firestore dependency from CLI (#224)
**Problem**

Current implementation of CLI depends on firestore access for
1. reading configuration
2. listen to request status changes.

This creates firestore stickiness and limits the deployment to gcloud.
We might need to strip away this dependency to move the datastore and
cloud provider in the backend.

**Change**
1. removes firestore connection from the CLI.
2. uses /orgs endpoint for fetching org data to understand login.
3. uses integration config endpoint to fetch configuration for each of
the resource integrations
4. uses request streaming option for requests that need to wait for the
request state to transition to resolved status (DONE, DENIED OR ERRORED
statuses)

**Validation**
1. updated all unit tests and snapshots
2. manual testing for SSH, request and ls.

**Tracking**

ENG-5460

**Attachments**




https://github.com/user-attachments/assets/9c785850-88c7-45f5-b641-800984f54575