ENG-5485 Update description of SSH args to explicitly specify options
Current text:
"Flags and positionals passed to the underlying ssh implementation."

New text:
"Options passed to the underlying ssh implementation."

We don't currently support specifying positionals after `--`.
Updated the description accordingly.

Specifically, the command to be executed is already an argument to `p0 ssh`
and must be specified before the `--`.