Friendly error message for errors from web application firewall (#243)
The WAF of the P0 backend returns error messages as an HTML page.
These error messages were not surfaced to the user, instead the CLI
failed parsing them as JSON and surfaced that error.

Changes:
- First check if the response is in the format that the WAF returns:
contains a <title> tag with the status code in it. If yes, return that
as an error message. If no, continue with parsing as JSON.
- Opportunistic change: rename `unauthenticatedFetch` -> `baseFetch` and `baseFetch` -> `authFetch`.