fix ssh failing to save host fingerprint (#249)
This fixes an issue where the P0 CLI is always overriding the
`UserKnownHostsFile` even if we don't find any host keys for ssh access.
Because the UserKnownHostFile doesn't exist we end up failing to save
the host keys after a user performs TOFU. This creates a poor user
experience since customers will now always need to approve the ssh
prompt.