Fix CLI double login bug, improve debug logging (#263)
On Windows, the NPM version check was broken, which led to the end user
sometimes needing to run the login command twice. This fixes the core
issues at play, but also improves debug logging for easier traceability.

- Adjusts the spawn of the NPM version check to account for the fact
that NPM isn't a .exe file.
- Adds a .on("error") check to the `exec` implementation to explicitly
handle errors (required to prevent uncaught exceptions)
- Adds global process error handlers to improve logging when uncaught
errors or unhandled rejections occur (without this change, nothing
printed to the terminal; with this change, it would have made it much
easier to debug this particular issue)
- Improves overall debug logging of the NPM version checking code paths
- Adjusts `.on("exit")` to `.on("close")` to allow stdio/stderr to
cleanly finish
- Bump NPM version check timeout from 1s to 2s to improve reliability
(it took generally longer on my local Virtualized Windows hardware -
could be either due to the Virtualization or Windows itself being
slower)

I tested this on both Windows and MacOS to test the login flows, debug
logging, and that SSH functionality still works