Implement `p0 print-bearer-token` (#289)
### What
Adds the `print-bearer-token` command to the CLI - this command takes no
arguments and simply prints the bearer token to stdout.

### Why
The token is a necessary input to certain terraform operations.
Current documentation recommends that users use other command line tools
(`jq` or `python`) to extract the token string from the `identity.json`
file located in the user's home directory.

Adding a command to the CLI to pull this value makes setup simpler, and
removes the need to employ additional command line tools (which may not
be available by default).

### How
This PR implements the new command in a file called
`print-bearer-token.ts` - the path to the identity file is constructed,
data is read from it, and the value of the `credentials.access_token` is
read from the parsed JSON.