ENG-6072 Improve Error Message for SCP Command (#292)
### What
This PR improves the error message that's printed when a user attempts
to include a username in their SCP source/destination argument. The
output includes a reminder that usernames are not required, and should
be omitted from the command (but only does so if "@" is included in the
instance name).

Before:
```
p0 scp ./test.txt gus_boling_p0_dev@dev-instance-gus-boling:/home/gus_boling_p0_dev/            
Could not find any instances matching 'gus_boling_p0_dev@dev-instance-gus-boling'
```

After :
```
p0 scp ./test.txt gus_boling_p0_dev@dev-instance-gus-boling:/home/gus_boling_p0_dev/            
Could not find any instances matching 'gus_boling_p0_dev@dev-instance-gus-boling'
Hint: The instance-name appears to contain a username AND a hostname; the username should be omitted.
```

### Why
The standard `scp` command requires users to specify both a username and
a hostname when moving files to/from a remote machine. The `p0 scp`
command implements the same functionality, but doesn't require a
username, since this is determined by the user's P0 identity.

This can be confusing because the syntax for both command is otherwise
very similar, and the existing error message could be interpreted as "p0
is unable to find a user's instance" rather than "p0 is unable to find
any instance starting with `<username>@...`".

### How
This PR modifies how errors are handled when making an SCP request; this
allows us to enhance the output for this specific use/error-case.