Wrap `az` commands in `osSafeCommand` function. (#303)
On Windows, the `az` CLI needs the `cmd.exe` utility to be properly
invoked programmatically since it is not an executable.

There were a number of places this was not being done and the P0 CLI
commands that use them were not working as a result.

This PR wraps every `az` command in the CLI code in a function
`osSafeCommand`. This wrapper function invokes commands using the
`cmd.exe` utility if it detects that the OS is Windows.

This function was adapted from the existing `azCommandArgs` function
which was already used to run `az` commands for this purpose.

To validate this worked, I ran `p0 ssh --provider azure` command from
the CLI on Windows (that `p0 CLI` command invokes all of the `az`
commands modified in this PR). After these changes, the ssh session
kicks off as expected. Prior to the change the `p0 CLI` would error out
when it tried to run the `az` commands.

More context:
https://github.com/p0-security/p0cli/pull/293#discussion_r2479743274