Remove ensureInstall check for windows ssh requests (#302)
The logic for requesting an ssh session for Azure performs a check to
ensure the necessary dependencies for the command are installed, in this
case the `az` CLI.

This runs into trouble because our current method for validating the
command works relies on Homebrew, which is not designed to work on
Windows.

Because of this behavior, the user will see an error telling them to
install Homebrew and their request will fail to provision.

<img width="885" height="316" alt="Screenshot 2025-11-07 at 2 10 08 PM"
src="https://github.com/user-attachments/assets/37410394-4644-421f-b458-70edc0b71beb"
/>

This PR removes the check on Windows for now so users will at least be
able to make requests.

There are follow-up items to add this functionality back for Windows.


https://linear.app/p0-security/issue/ENG-6471/rdp-beta-support-ensureinstall-check-on-windows-cli

https://linear.app/p0-security/issue/ENG-6472/rdp-beta-add-validation-az-cli-is-installed-on-the-users-machine