Use osSafeCommand for npm and gcloud CLI commands (#304)
A new function called `osSafeCommand` was introduced in
https://github.com/p0-security/p0cli/pull/303 for the purpose of running
shell commands appropriately on both Windows and Mac.

This PR updates two places in the CLI code where this check was already
being done to use the function instead. There is not functional change
here. `osSafeCommand` performs the same conditional logic that this code
was currently doing.'

I tested this by running Gcloud SSH commands on both Windows and Mac. I
confirmed that the version check and the provisioning of the SSH session
still work on both.