add support for isPreapproved to request (#324)
The backend returns a new flag `isPreapproved`, this can be used to make
the request less verbose if preapproval exists.

#### Before

```
./p0 ssh dev-instance
Access requested (see ... for details)
Will wait up to 5 minutes for this request to complete...
Your request was approved
Waiting for access to be provisioned
```

#### After

```
./p0 ssh dev-instance
Waiting for access to be provisioned
Authenticated to dev-instance (via proxy) using "publickey".
```