Print URL when using device code flow (#326)
Some users have reported that they have trouble logging in because they
are using headless machines that do not contain web browsers. For auth
flows that use the "device code" flow, the URL that the CLI opens can be
logged into via any device - it doesn't necessarily need to be the same
device that the CLI is running from.

We believe that simply printing the login URL and instructing users to
visit that URL manually is sufficient to unblock them. (This is purely a
UX change rather than a technical one).

I validated this by manually doing the p0 login and p0 ssh into an AWS
node, and completing the browser flow on a different device for both the
`p0 login` step and the AWS IDC assume-role login step